Wed, Sep 11, 2024, 4:42 PM +05:30

+++ credit/credit.c
@@ -5,7 +5,7 @@ int main(void)
 {
     long number = get_long("Enter the number\n");
 
-    long temp_number = number;
+    int temp_number = number;
     int length_count = 0;
 
     while(temp_number != 0)
